% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_id.R
\name{convert_id}
\alias{convert_id}
\title{Convert Gene IDs to Ensembl}
\usage{
convert_id(genes, data = NULL, species_id, id_type = "ens")
}
\arguments{
\item{genes}{A vector or character string of gene identifiers to convert.}

\item{data}{Optional data frame or matrix. If provided, the function attempts
to match \code{genes} to the row names or a column in \code{data} and merges the
conversion results with the original data.}

\item{species_id}{Numeric. The ID of the species for the database connection.}

\item{id_type}{Character. The type of ID to convert to:
\itemize{
\item "ens" = Ensembl gene IDs (Default)
\item "entrez" = Entrez gene IDs
\itemize{
\item WARNING: Not all species or genes have Entrez gene IDs available
\item May take longer than Ensembl IDs
\item Will likely have duplicate Entrez IDs
}
}}
}
\value{
A data frame.
\itemize{
\item If \code{data} is \code{NULL}: Returns a mapping table with original IDs and
IDs of selected type.
\item If \code{data} is provided: Returns \code{data} merged with the IDs of selected
type. Returns \code{NULL} if \code{species_id} is missing or no matches are found.
\item Any whitespace found in original IDs will be removed.
}
}
\description{
Queries the database to map user-provided gene identifiers to Ensembl IDs. To
ensure best matching and conversion, please verify that all gene identifiers
have no whitespace and are at least 2 characters long.
}
