% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{Get Table from Selected Database}
\usage{
get_table(species_id = NULL, table = NULL)
}
\arguments{
\item{species_id}{Numeric. The selected species ID.
If \code{NULL}, the function defaults to loading general organism info.}

\item{table}{Character. The name of the table to retrieve (e.g., "geneInfo", "pathway").
If \code{NULL}, defaults to "geneInfo" (if \code{species_id} is provided) or "orgInfo"
(if \code{species_id} is \code{NULL}).}
}
\value{
A data frame containing the data from the selected table.
}
\description{
Retrieves a specific table from the database for a selected species.
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) withAutoprint(\{ # examplesIf}
# Retrieve geneInfo table for Indian Cobra Species
cobra_genes <- get_table(species_id = 99,
                         table = "geneInfo")

# View table
head(cobra_genes)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link{list_tables}} to see available tables for a species.
}
