% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tables.R
\name{list_tables}
\alias{list_tables}
\title{List Table Options}
\usage{
list_tables(species_id = NULL)
}
\arguments{
\item{species_id}{Numeric. The selected species ID.
If \code{NULL}, loads the general organism info database.}
}
\value{
A character vector of table names available in the database connection.
}
\description{
Lists all available tables within the database for a specific species.
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) withAutoprint(\{ # examplesIf}
# List all tables available for species 99 (Indian Cobra)
list_tables(species_id = 99)
\dontshow{\}) # examplesIf}
}
