% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_categories.R
\name{path_categories}
\alias{path_categories}
\title{Retrieve Pathway Categories}
\usage{
path_categories(species_id = NULL)
}
\arguments{
\item{species_id}{Numeric. The ID of a desired species from database, found
using \code{srch_species()}}
}
\value{
Data frame of pathway categories for given species
}
\description{
Retrieves pathway category options (e.g. KEGG, GOBP, etc.) for a given
species. May take longer for well-documented species (i.e. Human)
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) withAutoprint(\{ # examplesIf}
# Get pathway categories for species 99 (Indian Cobra)
categories <- path_categories(species_id = 99)
head(categories)
\dontshow{\}) # examplesIf}
}
