% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathdb-package.R
\docType{package}
\name{pathdb-package}
\alias{pathdb}
\alias{pathdb-package}
\title{pathdb: Comprehensive Database for Pathway Enrichment Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides access to large-scale genomics data from the South Dakota State University's bioinformatics database, a unified platform for pathway analysis of over 13,000 organisms. It includes various gene mappings, gene characteristics, and pathway mapping data from KEGG, GOBP, GOCC, and many more pathway databases. Also provides various helper functions for processing RNA-Seq data for differential expression analysis and pathway enrichment analysis, occasionally sourced from code from iDEP, developed by Ge, S.X., Son, E.W. & Yao, R. (2018) \doi{10.1186/s12859-018-2486-6}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aidanfred24/pathdb}
  \item \url{https://aidanfred24.github.io/pathdb/}
  \item Report bugs at \url{https://github.com/aidanfred24/pathdb/issues}
}

}
\author{
\strong{Maintainer}: Aidan Frederick \email{Aidan.Frederick@sdstate.edu}

Other contributors:
\itemize{
  \item Xijin Ge \email{Xijin.Ge@sdstate.edu} [funder]
}

}
\keyword{internal}
