% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T2G_prep.R
\name{T2G_prep}
\alias{T2G_prep}
\title{TERM2GENE Data Prep for iDEP Database}
\usage{
T2G_prep(species_id = NULL, category = "GOBP", genes = NULL)
}
\arguments{
\item{species_id}{Numeric. The ID of a desired species from database, found
using \code{srch_species()}}

\item{category}{Character. A vector or character constant of pathway
categories/databases (e.g. KEGG, GOBP, GOCC, etc.). It is not recommended to
use all categories, as some species have many, leading to performance issues}

\item{genes}{Character. A character vector of genes to add to query}
}
\value{
A data frame containing TERM2GENE Data (pathways to genes)
}
\description{
Prepares background genes for enrichment analysis functions in the
format of TERM2GENE data, using pathway information from various databases.
Requires ID for a species, and can filter for specific vector of genes.
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) withAutoprint(\{ # examplesIf}
# CAUTION: The human database is very large, running these examples require
# the download of the human database.

# Prepare background genes mapping for Hypoxia dataset
# Useful for pathway enrichment analysis of our data
bg_genes <- T2G_prep(
  species_id = 96,
  category = "KEGG",
  genes = rownames(hypoxia_deseq)
)
head(bg_genes)
\dontshow{\}) # examplesIf}
}
